Replace KvFlag with what was previously KvFlag2
